fixes compilation warnings and UB for LP32 platforms
* convert pointers via uintptr_t to/from integral types
* mix in intermediate 64-bit `len` instead of platform-dependently sized `size_t`
* default to compiling with -D_FILE_OFFSET_BITS=64 for consistent sendfile64() usage

+++ lib/facil/redis/redis_engine.c
@@ -684,7 +684,7 @@ static void redis_on_publish_child(const fio_pubsub_engine_s *eng,
   fio_str_s tmp = FIO_STR_INIT;
   /* by using fio_str_s, short names are allocated on the stack */
   fio_str_info_s tmp_info = fio_str_resize(&tmp, channel.len + 8);
-  fio_u2str64(tmp_info.data, (uint64_t)eng);
+  fio_u2str64(tmp_info.data, (uintptr_t)eng);
   memcpy(tmp_info.data + 8, channel.data, channel.len);
   /* forward publication request to Root */
   fio_publish(.filter = -1, .channel = tmp_info, .message = msg,
@@ -701,7 +701,7 @@ Root Publication Handler
 static void redis_on_internal_publish(fio_msg_s *msg) {
   if (msg->channel.len < 8)
     return; /* internal error, unexpected data */
-  void *en = (void *)fio_str2u64(msg->channel.data);
+  void *en = (void *)(uintptr_t)fio_str2u64(msg->channel.data);
   if (en != msg->udata1)
     return; /* should be delivered by a different engine */
   /* step after the engine data */
@@ -721,8 +721,8 @@ Sending commands using the Root connection
 static void redis_forward_reply(fio_pubsub_engine_s *e, FIOBJ reply,
                                 void *udata) {
   uint8_t *data = udata;
-  fio_pubsub_engine_s *engine = (fio_pubsub_engine_s *)fio_str2u64(data + 0);
-  void *callback = (void *)fio_str2u64(data + 8);
+  fio_pubsub_engine_s *engine = (fio_pubsub_engine_s *)(uintptr_t)fio_str2u64(data + 0);
+  void *callback = (void *)(uintptr_t)fio_str2u64(data + 8);
   if (engine != e || !callback) {
     FIO_LOG_DEBUG("Redis reply not forwarded (callback: %p)", callback);
     return;
@@ -738,7 +738,7 @@ static void redis_forward_reply(fio_pubsub_engine_s *e, FIOBJ reply,
 static void redis_on_internal_cmd(fio_msg_s *msg) {
   // void*(void *)fio_str2u64(msg->msg.data);
   fio_pubsub_engine_s *engine =
-      (fio_pubsub_engine_s *)fio_str2u64(msg->channel.data + 0);
+      (fio_pubsub_engine_s *)(uintptr_t)fio_str2u64(msg->channel.data + 0);
   if (engine != msg->udata1) {
     return;
   }
@@ -756,7 +756,7 @@ static void redis_on_internal_cmd(fio_msg_s *msg) {
 /* Listens on filter `-10 -getpid()` for incoming reply data */
 static void redis_on_internal_reply(fio_msg_s *msg) {
   fio_pubsub_engine_s *engine =
-      (fio_pubsub_engine_s *)fio_str2u64(msg->channel.data + 0);
+      (fio_pubsub_engine_s *)(uintptr_t)fio_str2u64(msg->channel.data + 0);
   if (engine != msg->udata1) {
     FIO_LOG_DEBUG("Redis reply not forwarded (engine mismatch: %p != %p)",
                   (void *)engine, msg->udata1);
@@ -765,8 +765,8 @@ static void redis_on_internal_reply(fio_msg_s *msg) {
   FIOBJ reply;
   fiobj_json2obj(&reply, msg->msg.data, msg->msg.len);
   void (*callback)(fio_pubsub_engine_s *, FIOBJ, void *) = (void (*)(
-      fio_pubsub_engine_s *, FIOBJ, void *))fio_str2u64(msg->channel.data + 8);
-  void *udata = (void *)fio_str2u64(msg->channel.data + 16);
+      fio_pubsub_engine_s *, FIOBJ, void *))(uintptr_t)fio_str2u64(msg->channel.data + 8);
+  void *udata = (void *)(uintptr_t)fio_str2u64(msg->channel.data + 16);
   callback(engine, reply, udata);
   fiobj_free(reply);
 }
@@ -788,9 +788,9 @@ intptr_t redis_engine_send(fio_pubsub_engine_s *engine, FIOBJ command,
   fio_str_s tmp = FIO_STR_INIT;
   fio_str_info_s ti = fio_str_resize(&tmp, 28);
   /* combine metadata */
-  fio_u2str64(ti.data + 0, (uint64_t)engine);
-  fio_u2str64(ti.data + 8, (uint64_t)callback);
-  fio_u2str64(ti.data + 16, (uint64_t)udata);
+  fio_u2str64(ti.data + 0, (uintptr_t)engine);
+  fio_u2str64(ti.data + 8, (uintptr_t)callback);
+  fio_u2str64(ti.data + 16, (uintptr_t)udata);
   fio_u2str32(ti.data + 24, (uint32_t)getpid());
   FIOBJ cmd = fiobj2resp_tmp(command);
   fio_publish(.filter = -2, .channel = ti, .message = fiobj_obj2cstr(cmd),
